+++ src/main/java/com/koddy/server/global/utils/redis/StringRedisOperator.java
@@ -1,4 +1,4 @@
-package com.koddy.server.global.redis;
+package com.koddy.server.global.utils.redis;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.StringRedisTemplate;
