+++ src/test/java/com/koddy/server/auth/infrastructure/auth/RedisAuthenticationProcessorTest.java
@@ -3,8 +3,8 @@ package com.koddy.server.auth.infrastructure.auth;
 import com.koddy.server.auth.domain.model.code.AuthCodeGenerator;
 import com.koddy.server.auth.exception.AuthException;
 import com.koddy.server.common.RedisTest;
-import com.koddy.server.global.redis.RedisOperator;
-import com.koddy.server.global.redis.StringRedisOperator;
+import com.koddy.server.global.utils.redis.RedisOperator;
+import com.koddy.server.global.utils.redis.StringRedisOperator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
