+++ src/test/java/com/koddy/server/coffeechat/domain/model/StrategyTest.java
@@ -1,7 +1,7 @@
 package com.koddy.server.coffeechat.domain.model;
 
 import com.koddy.server.common.UnitTest;
-import com.koddy.server.global.encrypt.Encryptor;
+import com.koddy.server.global.utils.encrypt.Encryptor;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
