+++ src/test/java/com/koddy/server/common/utils/EncryptorFactory.java
@@ -1,7 +1,7 @@
 package com.koddy.server.common.utils;
 
 import com.koddy.server.common.mock.fake.FakeEncryptor;
-import com.koddy.server.global.encrypt.Encryptor;
+import com.koddy.server.global.utils.encrypt.Encryptor;
 
 import java.security.SecureRandom;
